[feature/#13-router] SimpleRouter implementiert
